# docs/openapi.yaml
openapi: 3.1.0
info:
  title: onepoint API
  version: 0.1.0
  description: |
    REST API for onepoint.

    ## Authentication
    Most endpoints require a self-managed **Bearer access token** (no JWT).
    1. `POST /api/register` to create an account.
    2. `POST /api/login` to obtain an `access_token` (+ `refresh_token`).
    3. Send `Authorization: Bearer <access_token>` on protected endpoints.
    4. `POST /api/token/refresh` to rotate an expired access token.
    5. `POST /api/logout` to revoke the current token.

    Access tokens live 1 hour; refresh tokens live 30 days. A user may hold
    many tokens at once (one per device); logging out revokes only the token
    presented.

    ### Testing here
    Use **Authorize** (top right) and paste the `access_token` from a login
    response to call protected endpoints with "Try it out".
servers:
  - url: http://localhost:8088
    description: Local Docker environment
...
paths:
...
  /api/admin/products/{id}/images/order:
    put:
      summary: Reorder a product's gallery
      description: Requires ROLE_ADMIN. `ids` must be the product's full set of image ids in the desired order; the first becomes primary.
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids: { type: array, items: { type: integer } }
      responses:
        '200':
          description: Reordered product
          content: { application/json: { schema: { $ref: '#/components/schemas/Product' } } }
        '403': { description: Not an admin }
        '404': { description: Product not found }
        '422': { description: ids is not the product's image set }
...
  /api/consultations:
    post:
      summary: Submit a design-consultation request
      description: Public storefront form submission.
      tags: [Catalog]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConsultationRequest' }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/Consultation' } } }
        '422':
          description: Validation failed
          content: { application/json: { schema: { $ref: '#/components/schemas/ValidationError' } } }
  /api/admin/orders:
    get:
      summary: List orders
      description: Requires ROLE_ADMIN. Newest first.
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending, confirmed, delivering, completed, cancelled] } }
        - { name: q, in: query, schema: { type: string }, description: Search order number / customer / email }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0 } }
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Order' } }
                  total: { type: integer }
        '403': { description: Not an admin }
        '422': { description: Invalid status filter }
  /api/admin/orders/{id}:
    get:
      summary: Get an order with line items
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Order detail
          content: { application/json: { schema: { $ref: '#/components/schemas/OrderDetail' } } }
        '403': { description: Not an admin }
        '404': { description: Order not found }
    patch:
      summary: Change an order's status
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderStatusRequest' }
      responses:
        '200':
          description: Updated
          content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } }
        '403': { description: Not an admin }
        '404': { description: Order not found }
        '422': { description: Invalid status }
...
  /api/admin/consultations:
    get:
      summary: List consultation requests
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [new, contacted, closed] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0 } }
      responses:
        '200':
          description: Consultations
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Consultation' } }
                  total: { type: integer }
        '403': { description: Not an admin }
        '422': { description: Invalid status filter }
  /api/admin/consultations/{id}:
    patch:
      summary: Change a consultation's status
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConsultationStatusRequest' }
      responses:
        '200':
          description: Updated
          content: { application/json: { schema: { $ref: '#/components/schemas/Consultation' } } }
        '403': { description: Not an admin }
        '404': { description: Consultation not found }
        '422': { description: Invalid status }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Paste the `access_token` returned by `/api/login`.
  schemas:
...
    ValidationError:
      description: Symfony validation error (RFC 7807-style).
      type: object
      properties:
        type:
          type: string
        title:
          type: string
          example: Validation Failed
        status:
          type: integer
          example: 422
        detail:
          type: string
        violations:
          type: array
          items:
            type: object
            properties:
              propertyPath:
                type: string
              title:
                type: string
    Product:
      type: object
      description: >
        A storefront product. `category` / `collection` are display names;
        `categorySlug` / `collectionSlug` are the matching URL slugs.
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        price: { type: integer, description: Whole manat }
        oldPrice: { type: [integer, "null"] }
        onSale: { type: boolean }
        isNew: { type: boolean }
        inStock: { type: boolean }
        freeDelivery: { type: boolean }
        warrantyYears: { type: integer }
        image: { type: [string, "null"], description: Primary gallery image (first), for back-compat with cards }
        images:
          type: array
          description: Ordered gallery; first is primary
          items: { $ref: '#/components/schemas/ProductImage' }
        category: { type: string }
        categoryId: { type: integer }
        categorySlug: { type: string }
        collection: { type: string }
        collectionId: { type: integer }
        collectionSlug: { type: string }
        material: { type: [string, "null"] }
        color: { type: [string, "null"] }
        dimensions: { type: [string, "null"] }
        description: { type: [string, "null"] }
    ProductImage:
      type: object
      properties:
        id: { type: integer }
        url: { type: string, description: Absolute URL or path relative to the uploads root }
        sortOrder: { type: integer, description: Present on the add/upload responses }
...
    Order:
      type: object
      properties:
        id: { type: integer, description: Numeric id for API calls }
        orderNumber: { type: string, example: OP-1042 }
        customer: { type: string }
        email: { type: string }
        items: { type: integer, description: Total item quantity }
        total: { type: integer, description: Whole manat }
        status: { type: string, enum: [pending, confirmed, delivering, completed, cancelled] }
        date: { type: string, format: date }
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          properties:
            lines:
              type: array
              items:
                type: object
                properties:
                  productId: { type: [integer, "null"] }
                  name: { type: string }
                  unitPrice: { type: integer }
                  quantity: { type: integer }
                  lineTotal: { type: integer }
    OrderStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, confirmed, delivering, completed, cancelled] }
...
    Consultation:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string }
        room: { type: string, enum: [living-room, bedroom, kitchen, kids, office] }
        message: { type: [string, "null"] }
        status: { type: string, enum: [new, contacted, closed] }
        createdAt: { type: string, format: date-time }
    ConsultationRequest:
      type: object
      required: [name, phone, room]
      properties:
        name: { type: string, maxLength: 120 }
        phone: { type: string, maxLength: 40 }
        room: { type: string, enum: [living-room, bedroom, kitchen, kids, office] }
        message: { type: [string, "null"], maxLength: 2000 }
    ConsultationStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [new, contacted, closed] }
...